requirements.txt
fastapi==0.115.4
pydantic==2.9.2
uvicorn==0.32.0
python-multipart==0.0.17
tensorflow==2.15.0
nltk==3.9.1
python-dotenv==1.0.1
requests==2.32.3